# docs/sample-api-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sample API",
    "version": "1.0.0",
    "description": "A sample API for testing documentation generation",
    "contact": {
      "name": "Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "summary": "Get users",
        "description": "GET /api/users - Handled by anonymous",
        "operationId": "get_api_users",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of items to return",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 10
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of items to skip",
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "users ID"
                      },
                      "name": {
                        "type": "string",
                        "description": "users name"
                      },
                      "description": {
                        "type": "string",
                        "description": "users description"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Creation timestamp"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Last update timestamp"
                      }
                    },
                    "required": [
                      "id",
                      "name"
                    ]
                  }
                },
                "example": [
                  {
                    "id": "123",
                    "name": "Example users",
                    "description": "This is an example users",
                    "createdAt": "2023-01-01T00:00:00Z",
                    "updatedAt": "2023-01-01T00:00:00Z"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Bad Request",
                  "message": "Invalid input parameters"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal Server Error",
                  "message": "An unexpected error occurred"
                }
              }
            }
          }
        },
        "tags": [
          "api"
        ]
      },
      "post": {
        "summary": "Create users",
        "description": "Middleware: authenticate",
        "operationId": "post_api_users",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "users ID"
                    },
                    "name": {
                      "type": "string",
                      "description": "users name"
                    },
                    "description": {
                      "type": "string",
                      "description": "users description"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Creation timestamp"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Last update timestamp"
                    }
                  },
                  "required": [
                    "id",
                    "name"
                  ]
                },
                "example": {
                  "id": "123",
                  "name": "Example users",
                  "description": "This is an example users",
                  "createdAt": "2023-01-01T00:00:00Z",
                  "updatedAt": "2023-01-01T00:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Bad Request",
                  "message": "Invalid input parameters"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal Server Error",
                  "message": "An unexpected error occurred"
                }
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "requestBody": {
          "description": "Request body for POST /api/users",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "users name"
                  },
                  "description": {
                    "type": "string",
                    "description": "users description"
                  }
                },
                "required": [
                  "name"
                ]
              },
              "example": {
                "name": "Example users",
                "description": "This is an example users"
              }
            }
          }
        }
      }
    },
...
  },
...
}